webpack now auto-recreates the entry files. docs updated with more info on webpack

+++ django_mako_plus/management/commands/dmp_webpack.py
@@ -1,12 +1,10 @@
 from django.apps import apps
-from django.utils.module_loading import import_string
 from django.conf import settings
 from django.core.management.base import BaseCommand, CommandError
 from django_mako_plus.template import create_mako_context
 from django_mako_plus.provider.runner import ProviderRun
 from django_mako_plus.management.mixins import DMPCommandMixIn
 from mako.template import Template as MakoTemplate
-
 import os
 import os.path
 from collections import OrderedDict
@@ -17,6 +15,11 @@ class Command(DMPCommandMixIn, BaseCommand):
     help = 'Removes compiled template cache folders in your DMP-enabled app directories.'
     can_import_settings = True
 
+    def __init__(self, *args, running_inline=False, **kwargs):
+        # special option when called from providers/webpack.py
+        self.running_inline = running_inline
+        super().__init__(*args, **kwargs)
+
 
     def add_arguments(self, parser):
         super().add_arguments(parser)
@@ -26,7 +29,7 @@ class Command(DMPCommandMixIn, BaseCommand):
             action='store_true',
             dest='overwrite',
             default=False,
-            help='Overwrite existing __entry__.js if necessary'
+            help='Overwrite existing __entry__.js if it exists'
         )
         parser.add_argument(
             '--single',
@@ -62,11 +65,12 @@ class Command(DMPCommandMixIn, BaseCommand):
             enapps = dmp.get_registered_apps()
 
         # main runner for per-app files
+        created = False
         if options.get('single') is None:
             for app in enapps:
                 self.message('Searching `{}` app...'.format(app.name))
                 filename = os.path.join(app.path, 'scripts', '__entry__.js')
-                self.create_entry_file(filename, self.generate_script_map(app), [ app ])
+                created = created or self.create_entry_file(filename, self.generate_script_map(app), [ app ])
 
         # main runner for one sitewide file
         else:
@@ -74,7 +78,7 @@ class Command(DMPCommandMixIn, BaseCommand):
             for app in enapps:
                 self.message('Searching `{}` app...'.format(app.name))
                 script_map.update(self.generate_script_map(app))
-            self.create_entry_file(options.get('single'), script_map, enapps)
+            created = created or self.create_entry_file(options.get('single'), script_map, enapps)
 
 
     def create_entry_file(self, filename, script_map, enapps):
@@ -82,25 +86,11 @@ class Command(DMPCommandMixIn, BaseCommand):
         if len(script_map) == 0:
             return
 
-        # delete previous file if it exists, and ensure the target directory is there
-        if os.path.exists(filename):
-            if self.options.get('overwrite'):
-                os.remove(filename)
-            else:
-                raise CommandError('Refusing to destroy existing file: {} (use --overwrite option or remove the file)'.format(filename))
-        if not os.path.exists(os.path.dirname(filename)):
-            os.makedirs(os.path.dirname(filename))
-
-        # write the entry file
-        self.message('Creating {}'.format(os.path.relpath(filename, settings.BASE_DIR)))
+        # create the entry file
         template = MakoTemplate('''
-<%! from datetime import datetime %>
-<%! import sys %>
 <%! import os %>
 
-// Generated on ${ datetime.now().strftime('%Y-%m-%d %H:%M') } by `${ ' '.join(sys.argv) }`
 // Contains links for ${ 'app' if len(enapps) == 1 else 'apps' }: ${ ', '.join(sorted([ a.name for a in enapps ])) }
-
 (context => {
     DMP_CONTEXT.loadBundle({
       %for (app, template), script_paths in script_map.items():
@@ -113,13 +103,33 @@ class Command(DMPCommandMixIn, BaseCommand):
     });
 })(DMP_CONTEXT.get());
 ''')
-        with open(filename, 'w') as fout:
-            fout.write(template.render(
-                enapps=enapps,
-                script_map=script_map,
-                filename=filename,
-            ).strip())
+        content = template.render(
+            enapps=enapps,
+            script_map=script_map,
+            filename=filename,
+        ).strip()
 
+        # ensure the parent directories exist
+        if not os.path.exists(os.path.dirname(filename)):
+            os.makedirs(os.path.dirname(filename))
+
+        # if the file exists, then consider the options
+        file_exists = os.path.exists(filename)
+        if file_exists and self.running_inline:
+            # running inline means that we're in debug mode and webpack is likely watching, so
+            # we don't want to recreate the entry file (and cause webpack to constantly reload)
+            # unless we have changes
+            with open(filename, 'r') as fin:
+                if content == fin.read():
+                    return False
+        if file_exists and not self.options.get('overwrite'):
+            raise CommandError('Refusing to destroy existing file: {} (use --overwrite option or remove the file)'.format(filename))
+
+        # if we get here, write the file
+        self.message('Creating {}'.format(os.path.relpath(filename, settings.BASE_DIR)))
+        with open(filename, 'w') as fout:
+            fout.write(content)
+        return True
 
     def generate_script_map(self, config):
         '''
